refactor: remove PathString and PathGraphql implementations
