+++ benches/request_template_bench.rs
@@ -2,9 +2,10 @@ use criterion::{black_box, criterion_group, criterion_main, Criterion};
 use derive_setters::Setters;
 use hyper::HeaderMap;
 use serde_json::json;
-use tailcall::{endpoint::Endpoint, path_value::PathValue};
+use tailcall::endpoint::Endpoint;
 use tailcall::has_headers::HasHeaders;
 use tailcall::http::RequestTemplate;
+use tailcall::path_value::PathValue;
 
 #[derive(Setters)]
 struct Context {
@@ -20,7 +21,8 @@ impl Default for Context {
 impl PathValue for Context {
     fn get_path_value<Path>(&self, path: &[Path]) -> Option<async_graphql::Value>
     where
-        Path: AsRef<str> {
+        Path: AsRef<str>,
+    {
         self.value.get_path_value(path)
     }
 }
