+++ src/lambda/evaluation_context.rs
@@ -8,6 +8,7 @@ use reqwest::header::HeaderMap;
 
 use super::{GraphQLOperationContext, ResolverContextLike};
 use crate::http::RequestContext;
+use crate::path_value::PathValue;
 
 // TODO: rename to ResolverContext
 #[derive(Clone)]
@@ -58,43 +59,25 @@ impl<'a, Ctx: ResolverContextLike<'a>> EvaluationContext<'a, Ctx> {
         self.graphql_ctx.value()
     }
 
-    pub fn path_arg<T: AsRef<str>>(&self, path: &[T]) -> Option<Cow<'a, Value>> {
+    pub fn path_arg<T: AsRef<str>>(&self, path: &[T]) -> Option<Value> {
         // TODO: add unit tests for this
-        if let Some(args) = self.graphql_ctx_args.as_ref() {
-            get_path_value(args.as_ref(), path).map(|a| Cow::Owned(a.clone()))
+        if let Some(v) = &self.graphql_ctx_args {
+            v.get_path_value(path)
         } else {
-            let arg = self.graphql_ctx.args()?.get(path[0].as_ref())?;
-            get_path_value(arg, &path[1..]).map(Cow::Borrowed)
+            self.graphql_ctx.args().and_then(|v| v.get_path_value(path))
         }
     }
 
-    pub fn path_value<T: AsRef<str>>(&self, path: &[T]) -> Option<Cow<'a, Value>> {
+    pub fn path_value<T: AsRef<str>>(&self, path: &[T]) -> Option<Value> {
         // TODO: add unit tests for this
-        if let Some(value) = self.graphql_ctx_value.as_ref() {
-            get_path_value(value.as_ref(), path).map(|a| Cow::Owned(a.clone()))
-        } else {
-            get_path_value(self.graphql_ctx.value()?, path).map(Cow::Borrowed)
-        }
+        self.graphql_ctx_value
+            .as_deref()
+            .or(self.graphql_ctx.value())
+            .and_then(|v| v.get_path_value(path))
     }
 
     pub fn headers(&self) -> &HeaderMap {
         &self.req_ctx.request_headers
-    }
-
-    pub fn header(&self, key: &str) -> Option<&str> {
-        let value = self.headers().get(key)?;
-
-        value.to_str().ok()
-    }
-
-    pub fn env_var(&self, key: &str) -> Option<String> {
-        self.req_ctx.runtime.env.get(key)
-    }
-
-    pub fn var(&self, key: &str) -> Option<&str> {
-        let vars = &self.req_ctx.server.vars;
-
-        vars.get(key).map(|v| v.as_str())
     }
 
     pub fn vars(&self) -> &BTreeMap<String, String> {
@@ -164,81 +147,225 @@ fn format_selection_field_arguments(field: SelectionField) -> Cow<'static, str> 
     Cow::Owned(format!("({})", args))
 }
 
-// TODO: this is the same code as src/json/json_like.rs::get_path
-pub fn get_path_value<'a, T: AsRef<str>>(input: &'a Value, path: &[T]) -> Option<&'a Value> {
-    let mut value = Some(input);
-    for name in path {
-        match value {
-            Some(Value::Object(map)) => {
-                value = map.get(name.as_ref());
-            }
-
-            Some(Value::List(list)) => {
-                value = list.get(name.as_ref().parse::<usize>().ok()?);
-            }
-            _ => return None,
-        }
+impl<'ctx, Ctx: ResolverContextLike<'ctx>> PathValue for EvaluationContext<'ctx, Ctx> {
+    fn get_path_value<Path>(&self, path: &[Path]) -> Option<async_graphql::Value>
+    where
+        Path: AsRef<str>,
+    {
+        path.split_first()
+            .and_then(|(head, tail)| match head.as_ref() {
+                "value" => self.path_value(tail),
+                "args" => self.path_arg(tail),
+                "headers" => self.headers().get_path_value(tail),
+                "vars" => self.vars().get_path_value(tail),
+                "env" => self.req_ctx.runtime.env.get_path_value(tail),
+                _ => None,
+            })
     }
-
-    value
 }
 
 #[cfg(test)]
 mod tests {
-    use async_graphql::Value;
-    use serde_json::json;
+    mod impl_path_value {
+        use std::collections::BTreeMap;
+        use std::sync::Arc;
 
-    use crate::lambda::evaluation_context::get_path_value;
+        use async_graphql::SelectionField;
+        use async_graphql_value::{ConstValue as Value, Name, Number};
+        use hyper::header::HeaderValue;
+        use hyper::HeaderMap;
+        use indexmap::IndexMap;
+        use once_cell::sync::Lazy;
 
-    #[test]
-    fn test_path_value() {
-        let json = json!(
-        {
-            "a": {
-                "b": {
-                    "c": "d"
-                }
+        use crate::http::RequestContext;
+        use crate::lambda::{EvaluationContext, ResolverContextLike};
+        use crate::path_value::PathValue;
+        use crate::EnvIO;
+
+        struct Env {
+            env: BTreeMap<String, String>,
+        }
+
+        impl EnvIO for Env {
+            fn get(&self, key: &str) -> Option<String> {
+                self.env.get(key).cloned()
             }
-        });
+        }
 
-        let async_value = Value::from_json(json).unwrap();
-
-        let path = vec!["a".to_string(), "b".to_string(), "c".to_string()];
-        let result = get_path_value(&async_value, &path);
-        assert!(result.is_some());
-        assert_eq!(result.unwrap(), &Value::String("d".to_string()));
-    }
-
-    #[test]
-    fn test_path_not_found() {
-        let json = json!(
-        {
-            "a": {
-                "b": "c"
+        impl Env {
+            pub fn init(map: BTreeMap<String, String>) -> Self {
+                Self { env: map }
             }
+        }
+
+        static TEST_VALUES: Lazy<Value> = Lazy::new(|| {
+            let mut root = IndexMap::new();
+            let mut nested = IndexMap::new();
+
+            nested.insert(
+                Name::new("existing"),
+                Value::String("nested-test".to_owned()),
+            );
+            root.insert(Name::new("bool"), Value::Boolean(true));
+            root.insert(Name::new("nested"), Value::Object(nested));
+            root.insert(Name::new("number"), Value::Number(Number::from(2)));
+            root.insert(Name::new("str"), Value::String("str-test".to_owned()));
+
+            Value::Object(root)
         });
 
-        let async_value = Value::from_json(json).unwrap();
+        static TEST_ARGS: Lazy<IndexMap<Name, Value>> = Lazy::new(|| {
+            let mut root = IndexMap::new();
+            let mut nested = IndexMap::new();
 
-        let path = vec!["a".to_string(), "b".to_string(), "c".to_string()];
-        let result = get_path_value(&async_value, &path);
-        assert!(result.is_none());
-    }
+            nested.insert(
+                Name::new("existing"),
+                Value::String("nested-test".to_owned()),
+            );
 
-    #[test]
-    fn test_numeric_path() {
-        let json = json!(
-        {
-            "a": [{
-                "b": "c"
-            }]
+            root.insert(Name::new("nested"), Value::Object(nested));
+            root.insert(Name::new("root"), Value::String("root-test".to_owned()));
+
+            root
         });
 
-        let async_value = Value::from_json(json).unwrap();
+        static TEST_HEADERS: Lazy<HeaderMap> = Lazy::new(|| {
+            let mut map = HeaderMap::new();
 
-        let path = vec!["a".to_string(), "0".to_string(), "b".to_string()];
-        let result = get_path_value(&async_value, &path);
-        assert!(result.is_some());
-        assert_eq!(result.unwrap(), &Value::String("c".to_string()));
+            map.insert("x-existing", HeaderValue::from_static("header"));
+
+            map
+        });
+
+        static TEST_VARS: Lazy<BTreeMap<String, String>> = Lazy::new(|| {
+            let mut map = BTreeMap::new();
+
+            map.insert("existing".to_owned(), "var".to_owned());
+
+            map
+        });
+
+        static TEST_ENV_VARS: Lazy<BTreeMap<String, String>> = Lazy::new(|| {
+            let mut map = BTreeMap::new();
+
+            map.insert("existing".to_owned(), "env".to_owned());
+
+            map
+        });
+
+        #[derive(Clone)]
+        struct MockGraphqlContext;
+
+        impl<'a> ResolverContextLike<'a> for MockGraphqlContext {
+            fn value(&'a self) -> Option<&'a Value> {
+                Some(&TEST_VALUES)
+            }
+
+            fn args(&'a self) -> Option<&'a IndexMap<Name, Value>> {
+                Some(&TEST_ARGS)
+            }
+
+            fn field(&'a self) -> Option<SelectionField> {
+                None
+            }
+
+            fn add_error(&'a self, _: async_graphql::ServerError) {}
+        }
+
+        static REQ_CTX: Lazy<RequestContext> = Lazy::new(|| {
+            let mut req_ctx = RequestContext::default().request_headers(TEST_HEADERS.clone());
+
+            req_ctx.server.vars = TEST_VARS.clone();
+            req_ctx.runtime.env = Arc::new(Env::init(TEST_ENV_VARS.clone()));
+
+            req_ctx
+        });
+
+        static EVAL_CTX: Lazy<EvaluationContext<'static, MockGraphqlContext>> =
+            Lazy::new(|| EvaluationContext::new(&REQ_CTX, &MockGraphqlContext));
+
+        #[test]
+        fn get_path() {
+            // value
+            assert_eq!(
+                EVAL_CTX.get_path_value(&["value", "bool"]),
+                Some(Value::Boolean(true))
+            );
+            assert_eq!(
+                EVAL_CTX.get_path_value(&["value", "number"]),
+                Some(Value::Number(Number::from(2)))
+            );
+            assert_eq!(
+                EVAL_CTX.get_path_value(&["value", "str"]),
+                Some(Value::String("str-test".to_owned()))
+            );
+            assert_eq!(
+                EVAL_CTX.get_path_value(&["value", "nested"]),
+                Some(serde_json::from_str(r#"{"existing":"nested-test"}"#).unwrap())
+            );
+            assert_eq!(EVAL_CTX.get_path_value(&["value", "missing"]), None);
+            assert_eq!(
+                EVAL_CTX.get_path_value(&["value", "nested", "missing"]),
+                None
+            );
+            assert_eq!(
+                    EVAL_CTX.get_path_value(&["value"]),
+                    Some(serde_json::from_str(r#"{"bool":true,"nested":{"existing":"nested-test"},"number":2,"str":"str-test"}"#).unwrap())
+                );
+
+            // args
+            assert_eq!(
+                EVAL_CTX.get_path_value(&["args", "root"]),
+                Some(Value::String("root-test".to_owned()))
+            );
+            assert_eq!(
+                EVAL_CTX.get_path_value(&["args", "nested"]),
+                Some(serde_json::from_str(r#"{"existing":"nested-test"}"#).unwrap())
+            );
+            assert_eq!(EVAL_CTX.get_path_value(&["args", "missing"]), None);
+            assert_eq!(
+                EVAL_CTX.get_path_value(&["args", "nested", "missing"]),
+                None
+            );
+            assert_eq!(
+                EVAL_CTX.get_path_value(&["args"]),
+                Some(
+                    serde_json::from_str(
+                        r#"{"nested":{"existing":"nested-test"},"root":"root-test"}"#
+                    )
+                    .unwrap()
+                )
+            );
+
+            // headers
+            assert_eq!(
+                EVAL_CTX.get_path_value(&["headers", "x-existing"]),
+                Some(Value::String("header".to_owned()))
+            );
+            assert_eq!(EVAL_CTX.get_path_value(&["headers", "x-missing"]), None);
+
+            // vars
+            assert_eq!(
+                EVAL_CTX.get_path_value(&["vars", "existing"]),
+                Some(Value::String("var".to_owned()))
+            );
+            assert_eq!(EVAL_CTX.get_path_value(&["vars", "missing"]), None);
+            assert_eq!(
+                EVAL_CTX.get_path_value(&["vars"]),
+                Some(serde_json::from_str(r#"{"existing":"var"}"#).unwrap())
+            );
+
+            // envs
+            assert_eq!(
+                EVAL_CTX.get_path_value(&["env", "existing"]),
+                Some(Value::String("env".to_owned()))
+            );
+            assert_eq!(EVAL_CTX.get_path_value(&["env", "x-missing"]), None);
+
+            // other value types
+            assert_eq!(EVAL_CTX.get_path_value(&["foo", "key"]), None);
+            assert_eq!(EVAL_CTX.get_path_value(&["bar", "key"]), None);
+            assert_eq!(EVAL_CTX.get_path_value(&["baz", "key"]), None);
+        }
     }
 }
