+++ src/serde_value_ext.rs
@@ -4,14 +4,15 @@ use anyhow::Result;
 use async_graphql::{Name, Value as GraphQLValue};
 use indexmap::IndexMap;
 
-use crate::{blueprint::DynamicValue, path_value::PathValue};
+use crate::blueprint::DynamicValue;
+use crate::path_value::PathValue;
 
 pub trait ValueExt {
     fn render_value(&self, ctx: &impl PathValue) -> Result<GraphQLValue>;
 }
 
 impl ValueExt for DynamicValue {
-    fn render_value<'a>(&self, ctx: &'a impl PathValue) -> Result<GraphQLValue> {
+    fn render_value(&self, ctx: &impl PathValue) -> Result<GraphQLValue> {
         match self {
             DynamicValue::Value(value) => Ok(value.to_owned()),
             DynamicValue::Mustache(m) => Ok(m.render(ctx).unwrap_or_default()),
